# src/swagger.json
{
	"components": {
...
		"schemas": {
			"Location": {
				"description": "Information about a location.",
				"properties": {
					"locationId": {
						"type": "string"
					},
					"address": {
						"type": "string"
					}
				},
				"required": [
					"locationId",
					"address"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"locationId": "location-1",
					"address": "Location 1"
				}
			},
			"SearchLocationsResult": {
				"description": "JSON response format for a search locations operation.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							200
						],
						"nullable": false
					},
					"data": {
						"properties": {
							"locations": {
								"items": {
									"$ref": "#/components/schemas/Location"
								},
								"type": "array"
							}
						},
						"required": [
							"locations"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": 200,
					"data": {
						"locations": [
							{
								"locationId": "location-1",
								"address": "Location 1"
							},
							{
								"locationId": "location-2",
								"address": "Location 2"
							}
						]
					}
				}
			},
			"ServerErrorResponse": {
				"description": "JSON response format for a \"500 Internal Server Error\" error.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							500
						],
						"nullable": false
					},
					"error": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					}
				},
				"required": [
					"status",
					"error"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": 500,
					"error": {}
				}
			},
			"CreateLocationResult": {
				"description": "JSON response format for a create location operation.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							201
						],
						"nullable": false
					},
					"data": {
						"properties": {
							"location": {
								"$ref": "#/components/schemas/Location"
							}
						},
						"required": [
							"location"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": 201,
					"data": {
						"created": {
							"locationId": "location-1",
							"address": "Location 1"
						}
					}
				}
			},
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"BadRequestErrorResponse": {
				"description": "JSON response format for a \"400 Bad Request\" error.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							400
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"fields": {
								"$ref": "#/components/schemas/FieldErrors"
							},
							"message": {
								"type": "string"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"status",
					"error"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": 400,
					"error": {
						"message": "Error description.",
						"fields": [
							{
								"name": {
									"message": "Reason or description.",
									"value": "Any value or undefined."
								}
							}
						]
					}
				}
			},
			"AuthenticationErrorResponse": {
				"description": "JSON response format for a \"401 Unauthorized\" error.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							401
						],
						"nullable": false
					},
					"error": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					}
				},
				"required": [
					"status",
					"error"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": 401,
					"error": {}
				}
			},
			"AuthorizationErrorResponse": {
				"description": "JSON response format for a \"403 Forbidden\" error.",
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							403
						],
						"nullable": false
					},
					"error": {
						"properties": {},
						"additionalProperties": {
							"type": "string"
						},
						"type": "object"
					}
				},
				"required": [
					"status",
					"error"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": 403,
					"error": {}
				}
			},
			"CreateLocationParams": {
				"description": "JSON request format to create a new location.",
				"properties": {
					"address": {
						"type": "string"
					}
				},
				"required": [
					"address"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"address": "Location 1"
				}
			},
			"ProductStatus": {
				"enum": [
					"available",
					"sold out",
					"last units",
					"no info"
				],
				"type": "string"
			},
			"ProductPublicInfo": {
				"properties": {
					"productId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"active": {
						"type": "boolean"
					},
					"status": {
						"$ref": "#/components/schemas/ProductStatus"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"category": {
						"type": "string"
					},
					"stock": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"productId",
					"name",
					"description",
					"price",
					"active",
					"status",
					"images",
					"category",
					"stock"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetProductsResult": {
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							200
						],
						"nullable": false
					},
					"data": {
						"properties": {
							"products": {
								"items": {
									"$ref": "#/components/schemas/ProductPublicInfo"
								},
								"type": "array"
							}
						},
						"required": [
							"products"
						],
						"type": "object"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductCategory": {
				"enum": [
					"tshirt",
					"book",
					"bag",
					"poster"
				],
				"type": "string"
			},
			"GetProductByIdResult": {
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							200
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/ProductPublicInfo"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PatchProductResult": {
				"properties": {
					"status": {
						"type": "number",
						"enum": [
							200
						],
						"nullable": false
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateProductParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"active": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"in": "cookie",
				"name": "auth_token"
			}
		}
	},
	"info": {
		"title": "IndieLisboa Stock Management",
		"version": "1.0.0",
		"description": "Web API to view/edit product stock, register sales and manage user accounts.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/locations": {
			"get": {
				"operationId": "GetLocations",
				"responses": {
					"200": {
						"description": "Successfully returned a list of locations.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SearchLocationsResult"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServerErrorResponse"
								}
							}
						}
					}
				},
				"summary": "Get a list of locations.",
				"tags": [
					"Locations"
				],
				"security": [
					{
						"jwt": [
							"manager"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "PostSale",
				"responses": {
					"201": {
						"description": "Successfully created a new location.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CreateLocationResult"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BadRequestErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": 400,
											"error": {
												"fields": {
													"body.address": {
														"message": "invalid string value",
														"value": 1
													}
												}
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthenticationErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthorizationErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServerErrorResponse"
								}
							}
						}
					}
				},
				"summary": "Create a new location.",
				"tags": [
					"Locations"
				],
				"security": [
					{
						"jwt": [
							"manager"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateLocationParams"
							}
						}
					}
				}
			}
		},
		"/products": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetProductsResult"
								}
							}
						}
					}
				},
				"description": "If a search criteria is applied, only products with an exact match will be returned.\r\nWhen applicable, if a search parameter has multiple values, returned products will match at least one of those values.",
				"summary": "Get a list of products. You may specify a search criteria.",
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"description": "Limit the number of products returned.",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						},
						"example": 10
					},
					{
						"description": "Used for pagination. When limit is used,\r\nchunks of products will be skipped (e.g. if page=5 and limit=10, the first 50 products will be skipped).",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 0
						},
						"example": 0
					},
					{
						"description": "Minimum product price (inclusive).",
						"in": "query",
						"name": "priceMin",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number",
							"minimum": 0
						},
						"example": 0
					},
					{
						"description": "Maximum product price (inclusive).",
						"in": "query",
						"name": "priceMax",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number",
							"maximum": 2000
						},
						"example": 50
					},
					{
						"description": "Product availability.",
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/components/schemas/ProductStatus"
							}
						}
					},
					{
						"description": "Product category.",
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ProductCategory"
						}
					}
				]
			}
		},
		"/products/{productId}": {
			"get": {
				"operationId": "GetProductById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetProductByIdResult"
								}
							}
						}
					}
				},
				"summary": "Get detailed information of a product.",
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "PatchProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PatchProductResult"
								}
							}
						}
					}
				},
				"summary": "Update a product's details.",
				"tags": [
					"Products"
				],
				"security": [
					{
						"jwt": [
							"manager"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProductParams"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}